homework3/requirements.txt
###### Requirements without Version Specifiers ######
psutil
PyYAML
flake8